# apps/backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/Courses": {
      "post": {
        "operationId": "createCourse",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseSingleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid course payload"
          },
          "409": {
            "description": "Course already exists"
          }
        },
        "summary": "Create a Course",
        "tags": [
          "Courses"
        ]
      }
    },
    "/api/Courses/getAll": {
      "post": {
        "operationId": "getCourses",
        "parameters": [
          {
            "name": "Degree",
            "required": false,
            "in": "query",
            "description": "Filter by Degree",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseFilters"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Courses returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseListResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get all courses",
        "tags": [
          "Courses"
        ]
      }
    },
    "/api/Courses/v2/getAll": {
      "get": {
        "description": "Use filters and enable stats mode in filters",
        "operationId": "getCoursesV2",
        "parameters": [
          {
            "name": "UniversityID",
            "required": false,
            "in": "query",
            "description": "Unique identifier for a university",
            "schema": {
              "format": "uuid",
              "example": "00000000-0000-0000-0000-000000000000",
              "type": "string"
            }
          },
          {
            "name": "CourseName",
            "required": false,
            "in": "query",
            "description": "Name of the course",
            "schema": {
              "minLength": 1,
              "maxLength": 255,
              "example": "Computer Science",
              "type": "string"
            }
          },
          {
            "name": "Degree",
            "required": false,
            "in": "query",
            "description": "Degree that course belongs to",
            "schema": {
              "minLength": 1,
              "maxLength": 30,
              "nullable": true,
              "example": "Bachelor of Science",
              "type": "string"
            }
          },
          {
            "name": "Stats",
            "required": false,
            "in": "query",
            "description": "Enable stats = TRUE",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Courses returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseListResponseDtoV2"
                }
              }
            }
          }
        },
        "summary": "Get all courses - V2",
        "tags": [
          "Courses"
        ]
      }
    },
    "/api/Courses/{CourseId}": {
      "get": {
        "operationId": "getCourseById",
        "parameters": [
          {
            "name": "CourseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseSingleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Course ID"
          },
          "404": {
            "description": "Course not found"
          }
        },
        "summary": "get a Course by ID",
        "tags": [
          "Courses"
        ]
      },
      "patch": {
        "operationId": "updateCourse",
        "parameters": [
          {
            "name": "CourseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseSingleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid update payload or CourseId"
          },
          "404": {
            "description": "Course not found"
          }
        },
        "summary": "Update an Course",
        "tags": [
          "Courses"
        ]
      },
      "delete": {
        "operationId": "deleteCourse",
        "parameters": [
          {
            "name": "CourseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCourseResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Course ID"
          },
          "404": {
            "description": "Course not found"
          }
        },
        "summary": "Delete Course by Course ID",
        "tags": [
          "Courses"
        ]
      }
    },
    "/api/Courses/v2/{CourseId}": {
      "get": {
        "operationId": "getCourseByIdV2",
        "parameters": [
          {
            "name": "CourseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseSingleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Course ID"
          },
          "404": {
            "description": "Course not found"
          }
        },
        "summary": "get a Course by ID - V2",
        "tags": [
          "Courses"
        ]
      }
    },
...
  },
  "info": {
    "title": "UMTAS API",
    "description": "University Management Timetable & Scheduling System API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://capstone-vigil.dns.net.za",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "umtas-session",
        "description": "Session cookie (set automatically by BetterAuth)"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AcademicSemester": {
        "type": "string",
        "enum": [
          "SEMESTER_1",
          "SEMESTER_2",
          "YEAR"
        ],
        "description": "Academic period in which this module is taught. Null or omitted means the whole academic year."
      },
      "StylingDto": {
        "type": "object",
        "properties": {
          "colour": {
            "type": "string",
            "example": "#3B82F6"
          }
        },
        "required": [
          "colour"
        ]
      },
...
      "CourseModuleDto": {
        "type": "object",
        "properties": {
          "CourseModuleID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "ID to identify Course Module entry"
          },
          "GroupModuleID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "ID to identify GroupModule entry metadata is for"
          },
          "CourseID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "ID to identify course owning this CourseModule"
          },
          "Core": {
            "type": "boolean",
            "example": true,
            "description": "Identifies wether or not module is a core module of the course"
          },
          "SemesterOfStudy": {
            "type": "string",
            "nullable": true,
            "example": "Semester 1",
            "description": "Identifies which part of the year the module takes part in for the course"
          },
          "YearOfStudy": {
            "nullable": true,
            "type": "number",
            "example": 1,
            "description": "Which year does the module belong to for the course"
          }
        },
        "required": [
          "CourseModuleID",
          "GroupModuleID",
          "CourseID"
        ]
      },
      "EventCriteriaDto": {
        "type": "object",
        "properties": {
          "eventSource": {
            "enum": [
              "university",
              "personal"
            ],
            "type": "string"
          },
          "date": {
            "type": "string",
            "example": "2026-02-17",
            "description": "Required when the event is not recurring."
          },
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "monday",
              "tuesday",
              "wednesday",
              "thursday",
              "friday",
              "saturday",
              "sunday"
            ],
            "description": "Required when the event is recurring."
          },
          "startTime": {
            "type": "string",
            "example": "08:30"
          },
          "endTime": {
            "type": "string",
            "example": "10:20"
          },
          "moduleId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "eventSource",
          "startTime",
          "endTime"
        ]
      },
      "VenueDto": {
        "type": "object",
        "properties": {
          "venueId": {
            "type": "string",
            "format": "uuid"
          },
          "venueName": {
            "type": "string"
          }
        },
        "required": [
          "venueId",
          "venueName"
        ]
      },
      "EventDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "eventName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "activityCode": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 10
          },
          "activityType": {
            "type": "string",
            "enum": [
              "lecture",
              "tutorial",
              "prac",
              "test",
              "exam"
            ]
          },
          "eventCriteria": {
            "$ref": "#/components/schemas/EventCriteriaDto"
          },
          "isRecurring": {
            "type": "boolean"
          },
          "validated": {
            "type": "boolean"
          },
          "venues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VenueDto"
            }
          }
        },
        "required": [
          "eventId",
          "eventCriteria"
        ]
      },
...
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "UniversityID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a university"
          },
          "GroupID": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a group of modules"
          },
          "CourseName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "example": "Computer Science",
            "description": "Name of the course"
          },
          "Degree": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 30,
            "example": "Bachelor of Science",
            "description": "Degree that course belongs to"
          },
          "ExternalID": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 255,
            "example": "12345",
            "description": "Refer to course on external API"
          }
        },
        "required": [
          "UniversityID",
          "CourseName"
        ]
      },
      "ModulesDto": {
        "type": "object",
        "properties": {
          "moduleID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a module"
          },
          "moduleCode": {
            "type": "string",
            "minLength": 2,
            "maxLength": 15,
            "example": "COS332",
            "description": "Module code used by the university"
          },
          "moduleName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Computer Networks",
            "description": "Name of the module"
          },
          "moduleDescription": {
            "type": "string",
            "nullable": true,
            "example": "Introduction to computer networking concepts",
            "description": "Short module description"
          },
          "semester": {
            "nullable": true,
            "example": "SEMESTER_1",
            "description": "Academic period in which this module is taught. Null or omitted means the whole academic year.",
            "allOf": [
              {
                "$ref": "#/components/schemas/AcademicSemester"
              }
            ]
          },
          "styling": {
            "nullable": true,
            "description": "Styling to be used for a Module",
            "example": {
              "colour": "FFFFF"
            },
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/StylingDto"
              }
            ]
          },
          "CourseModuleInfo": {
            "nullable": true,
            "description": "Metadata attached to module when owned by course",
            "example": {
              "CourseModuleID": "00000000-0000-0000-0000-000000000000",
              "GroupModuleID": "00000000-0000-0000-0000-000000000000",
              "CourseID": "00000000-0000-0000-0000-000000000000",
              "core": true,
              "SemesterOfStudy": "Semester 1",
              "YearOfStudy": 1
            },
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CourseModuleDto"
              }
            ]
          },
          "validated": {
            "type": "boolean",
            "example": true,
            "description": "Whether the module has been approved by a university admin"
          },
          "Events": {
            "nullable": true,
            "description": "List of events for the module",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventDto"
            }
          },
          "ExternalID": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 255,
            "example": "12345",
            "description": "Refer to module on external API"
          },
          "Enrolled": {
            "type": "boolean",
            "example": false,
            "description": "Is a user enrolled to this module"
          }
        },
        "required": [
          "moduleID",
          "moduleCode",
          "moduleName"
        ]
      },
      "CourseSingleResponseDto": {
        "type": "object",
        "properties": {
          "CourseID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a course"
          },
          "UniversityID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a university"
          },
          "GroupID": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a group of modules"
          },
          "CourseName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "example": "Computer Science",
            "description": "Name of the course"
          },
          "Degree": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 30,
            "example": "Bachelor of Science",
            "description": "Degree that course belongs to"
          },
          "Modules": {
            "description": "Modules for the course.",
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModulesDto"
            }
          },
          "ExternalID": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 255,
            "example": "12345",
            "description": "Refer to course on external API"
          }
        },
        "required": [
          "CourseID",
          "UniversityID",
          "CourseName"
        ]
      },
      "CourseFilters": {
        "type": "object",
        "properties": {
          "UniversityID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a university"
          },
          "CourseName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "example": "Computer Science",
            "description": "Name of the course"
          },
          "Degree": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 30,
            "example": "Bachelor of Science",
            "description": "Degree that course belongs to"
          }
        }
      },
      "CourseDto": {
        "type": "object",
        "properties": {
          "CourseID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a course"
          },
          "UniversityID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a university"
          },
          "GroupID": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a group of modules"
          },
          "CourseName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "example": "Computer Science",
            "description": "Name of the course"
          },
          "Degree": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 30,
            "example": "Bachelor of Science",
            "description": "Degree that course belongs to"
          },
          "Modules": {
            "description": "Modules for the course.",
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModulesDto"
            }
          },
          "ExternalID": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 255,
            "example": "12345",
            "description": "Refer to course on external API"
          }
        },
        "required": [
          "CourseID",
          "UniversityID",
          "CourseName"
        ]
      },
      "CourseListResponseDto": {
        "type": "object",
        "properties": {
          "courses": {
            "description": "List of courses",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseDto"
            }
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "courses"
        ]
      },
      "CourseListResponseDtoV2": {
        "type": "object",
        "properties": {
          "courses": {
            "description": "List of courses",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseDto"
            }
          },
          "count": {
            "type": "number",
            "example": 68,
            "description": "Count of Modules"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "courses"
        ]
      },
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "UniversityID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a university"
          },
          "GroupID": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a group of modules"
          },
          "CourseName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "example": "Computer Science",
            "description": "Name of the course"
          },
          "Degree": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 30,
            "example": "Bachelor of Science",
            "description": "Degree that course belongs to"
          },
          "Modules": {
            "description": "Modules for the course.",
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModulesDto"
            }
          },
          "ExternalID": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 255,
            "example": "12345",
            "description": "Refer to course on external API"
          }
        }
      },
      "DeleteCourseResponseDto": {
        "type": "object",
        "properties": {
          "CourseName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "example": "Computer Science",
            "description": "Name of the course"
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "CourseName",
          "success"
        ]
      },
...
    }
  }
}